+++ client/app/components/user profile/profile.module.css
@@ -2,9 +2,43 @@
     display: flex;
     flex-direction: column;
     align-items: center;
+    height: 100vh;
+    width: 20vw;
+    padding-top: 10rem;
+    background: rgb(255, 255, 255);
+    position: fixed;
+    box-shadow: -13px 16px 43px 0px rgba(0, 0, 0, 0.25);
 }
 
-.icon{
-    font-size: 2em !important;
+.pageFather{
+    display: flex;
+    width: 100%;
+    height: 100vh;
+    overflow: hidden;
+    background: black;
 }
 
+
+.iconsContainer{
+    display: flex;
+    align-items: center;
+    width: 1rem;
+}
+
+.iconAndText{
+    display: flex;
+    gap: 1rem;
+}
+
+.Buttons{
+    display: flex;
+    flex-direction: column;
+    gap: 2rem;
+    color: rgba(36, 39, 90, 1)
+}
+
+@media(max-width: 1022px) {
+    .textSideBar{
+        display: none;
+    }
+}
