+++ client/app/layout.tsx
@@ -3,6 +3,7 @@ import { Inter } from "next/font/google";
 import "./globals.css";
 import Navbar from "../components/Navbar/Navbar";
 import Footer from "../components/Footer/Footer";
+import UserProfile from "./components/user profile/profile";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -13,17 +14,19 @@ export const metadata: Metadata = {
 };
 
 export default function RootLayout({
-    children,
+  children,
 }: Readonly<{
-    children: React.ReactNode;
+  children: React.ReactNode;
 }>) {
-
   return (
     <html lang="en">
       <body className={inter.className}>
-        <Navbar />
+        <div>
+            <Navbar />
+            <UserProfile />
           {children}
-        <Footer />
+          <Footer />
+        </div>
       </body>
     </html>
   );
