+++ client/app/pages/user/layout.module.css
@@ -2,6 +2,23 @@
     display: flex;
 }
 
-.container main {
-    flex-grow: 1;
+.container{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    position: fixed;
+    height: 100vh;
+    width: 15vw;
+    top: 0;
+    background: rgb(114, 80, 80);
+    box-shadow: -13px 16px 43px 0px rgba(0, 0, 0, 0.25);
+}
+
+
+.pageFather{
+    display: flex;
+    width: 100%;
+    height: 100vh;
+    overflow: hidden;
+    background: black;
 }
