+++ client/app/pages/user/myProfile.module.css
@@ -1,0 +1,25 @@
+.gridColumns{
+    box-sizing: border-box;
+    margin: 0;
+    display: grid;
+    position: relative;
+    top: 0;
+    width: 100%;
+    height: 100vh;
+    grid-template-columns: 0.176fr 1fr;
+    grid-template-rows: 0.15fr 1fr;
+    z-index: -10;
+}
+
+.cardMyProfile{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+    width: 100%;
+    grid-row-start: 2;
+    grid-row-end: 3;
+    grid-column-start: 2;
+    grid-column-end: 3;
+}
