--- client/app/pages/user/myprofile/page.tsx
@@ -1,17 +1,0 @@
-'use client'
-import RootLayout from "../layout"
-
-const Profile = () => {
-    return (
-        <div>
-            <h1>Mi Perfil</h1>
-            <h2>Mis datos</h2>
-            <h2>Seguridad</h2>
-            <h2>Tarjetas</h2>
-            <h2>Privacidad</h2>
-            <h2>Terminos y Condiciones</h2>
-        </div>
-    )
-}
-
-export default Profile
